Added database functionality
