+++ AMOC_PROJECT/templates/register.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Fancy</title>
+</head>
+<body>
+    <form id="login-form" action="/register" method="post">
+        <input type="text" id="email" name="email">
+        <input type="password" id="password" name="password">
+        <input type="submit" id="submit" name="submit">
+    </form>
+</body>
+</html>
