+++ amoc_database/src/templates/login.html
@@ -5,10 +5,6 @@
     <title>Fancy</title>
 </head>
 <body>
-    <form id="login-form" action="/login" method="post">
-        <input type="text" id="email" name="email">
-        <input type="password" id="password" name="password">
-        <input type="submit" id="submit" name="submit">
-    </form>
+    <p> {{ email }}</p>
 </body>
 </html>
