--- amoc_database/src/templates/register.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Fancy</title>
-</head>
-<body>
-    <form id="login-form" action="/reg" method="post">
-        <input type="text" id="email" name="email">
-        <input type="password" id="password" name="password">
-        <input type="submit" id="submit" name="submit">
-    </form>
-</body>
-</html>
